+++ Amazon.Pay.API.SDK/WebStore/CheckoutSession/CheckoutSessionResponse.cs
@@ -16,6 +16,7 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
             Buyer = new Buyer();
             ProviderMetadata = new ProviderMetadata();
             ShippingAddress = new Address();
+            BillingAddress = new Address();
             PaymentPreferences = new List<PaymentPreferences>();
             Constraints = new List<Constraint>();
         }
@@ -35,7 +36,7 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         /// <summary>
         /// URLs associated to the Checkout Session used for completing checkout
         /// </summary>
-        [JsonProperty(PropertyName = "webCheckoutDetail")]
+        [JsonProperty(PropertyName = "webCheckoutDetails")]
         public WebCheckoutDetails WebCheckoutDetails { get; internal set; }
 
         /// <summary>
@@ -47,7 +48,7 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         /// <summary>
         /// Payment details specified by the merchant, such as the amount and method for charging the buyer
         /// </summary>
-        [JsonProperty(PropertyName = "paymentDetail")]
+        [JsonProperty(PropertyName = "paymentDetails")]
         public PaymentDetails PaymentDetails { get; internal set; }
 
         /// <summary>
@@ -65,7 +66,7 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         /// <summary>
         /// State of the Checkout Session object.
         /// </summary>
-        [JsonProperty(PropertyName = "statusDetail")]
+        [JsonProperty(PropertyName = "statusDetails")]
         public StatusDetails StatusDetails { get; internal set; }
 
         /// <summary>
@@ -91,6 +92,12 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         /// </summary>
         [JsonProperty(PropertyName = "shippingAddress")]
         public Address ShippingAddress { get; internal set; }
+
+        /// <summary>
+        /// Billing address for buyer-selected payment instrument. Billing address is only available in EU or for PayOnly product type.
+        /// </summary>
+        [JsonProperty(PropertyName = "billingAddress")]
+        public Address BillingAddress { get; internal set; }
 
         /// <summary>
         /// List of payment instruments selected by the buyer.